fix bug 13650
Fail Log 
<Text>Failure. Expected result: GL_INVALID_VALUE got: GL_NO_ERROR. Object target: 0x00008d41. Source target: GL_TEXTURE_2D. Destination target: GL_TEXTURE_2D</Text>  

Root cause is the test try to create two render buffer object and try to call glcopyImageSubData with target set as normal texture. The test expect there is no normal texture bearing the same name as the two render buffer object and it shall fail the copy. 
But as the test did not clean it's state for each test. The texture with the same name actually exist and not deleted and the copy pass.  

Clean up the state (all the resources, textures, buffers, shaders, programs etc.) per test is needed.